Limit how many records are pulled from pg_stat_statements.
There's no need to send a metric if no calls of a query have occurred since the last check. So cache the number of calls between check runs and use that to determine what has been executed or not. Only query full metrics data for queries that have been executed.